# class-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Class Operations API
  version: 1.5.0
  license:
    name: Proprietary
    identifier: LicenseRef-ClassOps-Proprietary
  description: |
    API cho các màn React hiện có: xác thực, lập lịch lớp và vận hành giáo viên.
    JWT có `accountScope=PLATFORM|TENANT`; tenant chỉ được suy ra cho phiên TENANT.
    API tạo lớp/ca/override không nhận link online theo DEC-046.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /platform/tenants:
    get:
      operationId: listPlatformTenants
      summary: Tìm, lọc và phân trang tenant
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Search"}, {$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {name: status, in: query, schema: {type: string, enum: [ACTIVE, LOCKED]}}]
      responses: {"200": {description: Danh sách 1-based tối đa 100}}
    post:
      operationId: createPlatformTenant
      summary: Tạo nguyên tử tenant và Admin ban đầu
      tags: [Platform administration]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateTenant"}}}}
      responses: {"201": {description: Tenant + Admin ban đầu + mật khẩu 123456}, "409": {$ref: "#/components/responses/Conflict"}}
  /platform/tenants/{id}:
    get:
      operationId: getPlatformTenant
      summary: Xem tenant và Admin ban đầu
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses: {"200": {description: Chi tiết tenant}, "404": {$ref: "#/components/responses/NotFound"}}
    patch:
      operationId: renamePlatformTenant
      summary: Đổi tên tenant với optimistic version
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name, version], properties: {name: {type: string}, version: {type: integer}}}}}}
      responses: {"200": {description: Tenant đã cập nhật}, "409": {$ref: "#/components/responses/Conflict"}}
  /platform/tenants/{id}/status:
    patch:
      operationId: changePlatformTenantStatus
      summary: Khóa hoặc mở tenant
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {$ref: "#/components/requestBodies/StatusChange"}
      responses: {"200": {description: ACTIVE hoặc LOCKED}, "409": {$ref: "#/components/responses/Conflict"}}
  /platform/tenants/{id}/initial-administrator/status:
    patch:
      operationId: changeInitialAdministratorStatus
      summary: Khóa hoặc mở Admin ban đầu
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {$ref: "#/components/requestBodies/StatusChange"}
      responses: {"200": {description: Admin ban đầu đã cập nhật}}
  /platform/tenants/{id}/initial-administrator/credential-resets:
    post:
      operationId: resetInitialAdministratorCredential
      summary: Reset credential Admin ban đầu
      tags: [Platform administration]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {$ref: "#/components/requestBodies/CredentialReset"}
      responses: {"200": {description: "Mật khẩu 123456, READY và token cũ bị thu hồi"}, "409": {$ref: "#/components/responses/Conflict"}}
...
  /tenants/{slug}:
    get:
      operationId: getTenant
      tags: [Authentication]
      security: []
      summary: Lấy thông tin tenant trước đăng nhập
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: Tenant tồn tại
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Tenant"}
        "404": {$ref: "#/components/responses/NotFound"}
        "423": {$ref: "#/components/responses/Locked"}
...
  /classes/{classId}/hourly-rates:
    get:
      operationId: listClassHourlyRates
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/ClassId"}]
      responses: {"200": {description: Lịch sử đơn giá, khoảng kết thúc suy từ mức kế tiếp, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/HourlyRate"}}}}}}
    post:
      operationId: createClassHourlyRate
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/ClassId"}, {$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/HourlyRateInput"}}}}
      responses: {"201": {description: Đơn giá mới và các accrual bị ảnh hưởng đã reconcile, content: {application/json: {schema: {$ref: "#/components/schemas/HourlyRate"}}}}, "409": {$ref: "#/components/responses/Conflict"}}
  /classes/{classId}/hourly-rates/{rateId}:
    patch:
      operationId: updateClassHourlyRate
      tags: [Salary and payroll]
      parameters:
        - {$ref: "#/components/parameters/ClassId"}
        - {name: rateId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/HourlyRateUpdate"}}}}
      responses: {"200": {description: Đơn giá và accrual đã tính lại}, "409": {$ref: "#/components/responses/Conflict"}}
  /sessions/{sessionId}/completion-correction:
    patch:
      operationId: correctCompletedSession
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/SessionId"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CompletionCorrection"}}}}
      responses: {"200": {description: Buổi và salary revision mới}, "400": {$ref: "#/components/responses/ValidationError"}, "409": {$ref: "#/components/responses/Conflict"}}
  /salary/payroll:
    get:
      operationId: listSalaryPayroll
      tags: [Salary and payroll]
      parameters:
        - {$ref: "#/components/parameters/SalaryMonth"}
        - {$ref: "#/components/parameters/Search"}
        - {name: status, in: query, schema: {type: string, enum: [OWED, SETTLED, OVERPAID]}}
        - {$ref: "#/components/parameters/Page"}
        - {$ref: "#/components/parameters/PageSize"}
        - {name: sort, in: query, schema: {type: string, default: teacherName}}
      responses: {"200": {description: KPI và danh sách giáo viên, content: {application/json: {schema: {$ref: "#/components/schemas/PayrollPage"}}}}, "403": {$ref: "#/components/responses/Forbidden"}}
  /salary/payroll/{teacherId}:
    get:
      operationId: getTeacherPayroll
      tags: [Salary and payroll]
      parameters:
        - {$ref: "#/components/parameters/TeacherId"}
        - {$ref: "#/components/parameters/SalaryMonth"}
      responses: {"200": {description: Chi tiết accrual, điều chỉnh và thanh toán, content: {application/json: {schema: {$ref: "#/components/schemas/TeacherPayrollDetail"}}}}}
  /teachers/me/salary:
    get:
      operationId: getOwnSalary
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/SalaryMonth"}]
      responses: {"200": {description: Lương của giáo viên đang đăng nhập, content: {application/json: {schema: {$ref: "#/components/schemas/TeacherPayrollDetail"}}}}, "403": {$ref: "#/components/responses/Forbidden"}}
  /salary/adjustments:
    post:
      operationId: createSalaryAdjustment
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SalaryAdjustmentInput"}}}}
      responses: {"201": {description: Điều chỉnh đã ghi sổ}, "409": {$ref: "#/components/responses/Conflict"}}
  /salary/adjustments/{adjustmentId}:
    patch:
      operationId: updateSalaryAdjustment
      tags: [Salary and payroll]
      parameters: [{name: adjustmentId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SalaryAdjustmentUpdate"}}}}
      responses: {"200": {description: Điều chỉnh đã sửa và audit before/after}, "409": {$ref: "#/components/responses/Conflict"}}
  /salary/payments:
    post:
      operationId: createSalaryPayment
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SalaryPaymentInput"}}}}
      responses: {"201": {description: Thanh toán đã ghi sổ}, "409": {$ref: "#/components/responses/Conflict"}}
  /salary/payments/{paymentId}:
    patch:
      operationId: updateSalaryPayment
      tags: [Salary and payroll]
      parameters: [{name: paymentId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SalaryPaymentUpdate"}}}}
      responses: {"200": {description: Thanh toán đã sửa và audit before/after}, "409": {$ref: "#/components/responses/Conflict"}}
  /salary/payroll/export:
    get:
      operationId: exportSalaryPayroll
      tags: [Salary and payroll]
      parameters: [{$ref: "#/components/parameters/SalaryMonth"}, {$ref: "#/components/parameters/Search"}]
      responses:
        "200":
          description: Workbook tổng hợp và chi tiết theo bộ lọc
          content: {application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {schema: {type: string, format: binary}}}
  /finance/salary-summary:
    get:
      operationId: getSalaryFinanceSummary
      tags: [Salary and payroll]
      parameters: [{name: year, in: query, required: true, schema: {type: integer, minimum: 2000, maximum: 2200}}]
      responses: {"200": {description: Series 12 tháng gồm kỳ lương và dòng tiền ngày trả}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    TenantSlug: {name: slug, in: path, required: true, schema: {type: string}}
    ClassId: {name: classId, in: path, required: true, schema: {type: string, format: uuid}}
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
    TeacherId: {name: teacherId, in: path, required: true, schema: {type: string, format: uuid}}
    SalaryMonth: {name: month, in: query, required: true, schema: {type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$"}}
...
    Search: {name: search, in: query, schema: {type: string, default: ""}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 150}
  requestBodies:
    StatusChange:
      required: true
      content: {application/json: {schema: {type: object, required: [status, reason, version], properties: {status: {type: string, enum: [ACTIVE, LOCKED]}, reason: {type: string}, version: {type: integer, minimum: 0}}}}}
    CredentialReset:
      required: true
      content: {application/json: {schema: {type: object, required: [reason, version], properties: {reason: {type: string}, version: {type: integer, minimum: 0}}}}}
  responses:
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Forbidden:
      description: Không có quyền
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NotFound:
      description: Không tìm thấy tài nguyên trong tenant
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Conflict:
      description: |
        Preview stale, TEACHER_OVERLAP, ROOM_OVERLAP, STUDENT_OVERLAP,
        optimistic lock hoặc idempotency conflict.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Locked:
      description: Tenant hoặc tài khoản đã bị khóa mềm
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  schemas:
...
    CreateTenant:
      type: object
      required: [name, slug, initialAdmin]
      properties:
        name: {type: string}
        slug: {type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"}
        initialAdmin: {type: object, required: [username, displayName], properties: {username: {type: string}, displayName: {type: string}, email: {type: string, format: email}}}
...
    Tenant:
      type: object
      required: [id, slug, name, status]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        status: {type: string, enum: [ACTIVE, LOCKED]}
...
    HourlyRateInput:
      type: object
      required: [effectiveDate, hourlyRate, reason]
      properties:
        effectiveDate: {type: string, format: date}
        hourlyRate: {type: number, exclusiveMinimum: 0}
        reason: {type: string, minLength: 1, maxLength: 1000}
    HourlyRateUpdate:
      allOf:
        - {$ref: "#/components/schemas/HourlyRateInput"}
        - type: object
          required: [version]
          properties: {version: {type: integer, format: int64, minimum: 0}}
    HourlyRate:
      type: object
      required: [id, classId, effectiveDate, hourlyRate, version, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        classId: {type: string, format: uuid}
        effectiveDate: {type: string, format: date}
        effectiveTo: {type: [string, "null"], format: date}
        hourlyRate: {type: number, exclusiveMinimum: 0}
        version: {type: integer, format: int64, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    CompletionCorrection:
      type: object
      required: [startAt, endAt, actualTeacherId, reason, version]
      properties:
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        actualTeacherId: {type: string, format: uuid}
        reason: {type: string, minLength: 1, maxLength: 1000}
        version: {type: integer, format: int64, minimum: 0}
    SalaryAdjustmentInput:
      type: object
      required: [teacherId, salaryMonth, amount, reason]
      properties:
        teacherId: {type: string, format: uuid}
        salaryMonth: {type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$"}
        amount: {type: number, not: {const: 0}}
        reason: {type: string, minLength: 1, maxLength: 1000}
    SalaryAdjustmentUpdate:
      allOf:
        - {$ref: "#/components/schemas/SalaryAdjustmentInput"}
        - type: object
          required: [editReason, version]
          properties:
            editReason: {type: string, minLength: 1, maxLength: 1000}
            version: {type: integer, format: int64, minimum: 0}
    SalaryPaymentInput:
      type: object
      required: [teacherId, salaryMonth, paidAt, amount, method, confirmOverpayment]
      properties:
        teacherId: {type: string, format: uuid}
        salaryMonth: {type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$"}
        paidAt: {type: string, format: date}
        amount: {type: number, exclusiveMinimum: 0}
        method: {type: string, enum: [CASH, BANK_TRANSFER]}
        reference: {type: [string, "null"], maxLength: 250}
        confirmOverpayment: {type: boolean}
        overpaymentReason: {type: [string, "null"], maxLength: 1000}
    SalaryPaymentUpdate:
      allOf:
        - {$ref: "#/components/schemas/SalaryPaymentInput"}
        - type: object
          required: [editReason, version]
          properties:
            editReason: {type: string, minLength: 1, maxLength: 1000}
            version: {type: integer, format: int64, minimum: 0}
    PayrollMetrics:
      type: object
      required: [sessionCount, totalMinutes, accrued, adjustments, due, paid, outstanding]
      properties:
        sessionCount: {type: integer, format: int64, minimum: 0}
        totalMinutes: {type: integer, format: int64, minimum: 0}
        accrued: {type: number}
        adjustments: {type: number}
        due: {type: number}
        paid: {type: number}
        outstanding: {type: number}
        status: {type: string, enum: [OWED, SETTLED, OVERPAID]}
    PayrollTeacherRow:
      allOf:
        - {$ref: "#/components/schemas/PayrollMetrics"}
        - type: object
          required: [teacherId, teacherName]
          properties:
            teacherId: {type: string, format: uuid}
            teacherName: {type: string}
    PayrollPage:
      type: object
      required: [month, metrics, teachers]
      properties:
        month: {type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$"}
        metrics: {$ref: "#/components/schemas/PayrollMetrics"}
        teachers:
          allOf:
            - {$ref: "#/components/schemas/Page"}
            - type: object
              properties: {items: {type: array, items: {$ref: "#/components/schemas/PayrollTeacherRow"}}}
    TeacherPayrollDetail:
      type: object
      required: [month, teacherId, teacherName, metrics, accruals, adjustments, payments]
      properties:
        month: {type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$"}
        teacherId: {type: string, format: uuid}
        teacherName: {type: string}
        metrics: {$ref: "#/components/schemas/PayrollMetrics"}
        accruals: {type: array, items: {type: object}}
        adjustments: {type: array, items: {type: object}}
        payments: {type: array, items: {type: object}}
    Page:
      type: object
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items: {type: array, items: {}}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}
    Problem:
      type: object
      required: [type, title, status, detail, code, message, timestamp, traceId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        message: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        details: {type: object, additionalProperties: true}
        retryAfterSeconds: {type: integer, minimum: 1}
        timestamp: {type: string, format: date-time}
        traceId: {type: string}